Restructured to add Cluster Hunter
Added cluster hunter. As we now have more than one major plugin, I've
reorganised the directories so there's a "Plugins" root for these. Each
major plugin also has a handbook-and-practicals directory, including
test data, which are copied into a handbooks-and-practicals directory at
the root of the distribution copy. Have updated the bat and sh files
accordingly.